upd : Методы для затеменения вынесеныы в отдельный класс сцены, добавлены комментарии
